christmas updates.

+++ database/migrations/2025_12_19_091754_add_price_to_courses_table.php
@@ -1,10 +1,9 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddPriceToCoursesTable extends Migration
 {
     /**
      * Run the migrations.
@@ -12,7 +11,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('courses', function (Blueprint $table) {
-            //
+            if (!Schema::hasColumn('courses', 'price')) {
+                $table->decimal('price', 8, 2)->default(0.00)->after('status');
+            }
         });
     }
 
@@ -22,7 +23,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('courses', function (Blueprint $table) {
-            $table->dropColumn('price');
+            if (Schema::hasColumn('courses', 'price')) {
+                $table->dropColumn('price');
+            }
         });
     }
-};
+}
